+++ src/lib.rs
@@ -65,13 +65,11 @@ macro_rules! matches {
 }
 
 
-mod bytestream;
 mod error;
 mod stream;
 mod strspan;
 mod xmlchar;
 
-pub use bytestream::*;
 pub use error::*;
 pub use stream::*;
 pub use strspan::*;