+++ src/lib.rs
@@ -46,7 +46,7 @@ If you are looking for a more high-level solution - checkout
 
 #![cfg_attr(feature = "cargo-clippy", allow(unreadable_literal))]
 
-#![doc(html_root_url = "https://docs.rs/xmlparser/0.9.0")]
+#![doc(html_root_url = "https://docs.rs/xmlparser/0.10.0")]
 
 #![forbid(unsafe_code)]
 #![warn(missing_docs)]