+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.10.0] - 2019-09-14
 ### Changed
+- 10-15% faster parsing.
 - Merge `ByteStream` and `Stream`.
 - `StreamError::InvalidChar` signature changed.
 - `StreamError::InvalidChar` was split into `InvalidChar` and `InvalidCharMultiple`.
@@ -14,7 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Check for [NameStartChar](https://www.w3.org/TR/xml/#NT-NameStartChar)
   during qualified name parsing.
 
-  E.g. `<-p>` is invalid from now.
+  E.g. `<-p>` is an invalid tag name from now.
 - Qualified name with multiple `:` is an error now.
 - `]>` is a valid text/`CharData` now. Previously it was parsed as `DoctypeEnd`.
 
@@ -155,7 +158,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - `log` 0.3 -> 0.4
 
-[Unreleased]: https://github.com/RazrFalcon/xmlparser/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/RazrFalcon/xmlparser/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/RazrFalcon/xmlparser/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.7.0...v0.8.0